requirements.txt
# Python >= 3.9 required
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
